Refactor Tokenlon adapters to use SimpleAdapter
